fix GetAddrConvertTable bug，return all offsets now

+++ src/Neo.Compiler.MSIL/Program.cs
@@ -165,12 +165,12 @@ namespace Neo.Compiler
 
                 if (options.Optimize)
                 {
-                    var optimize = NefOptimizeTool.Optimize(bytes, out Dictionary<uint, uint> addrConvTable);
+                    var optimize = NefOptimizeTool.Optimize(bytes, out Dictionary<int, int> addrConvTable);
                     log.Log("optimization succ " + (((bytes.Length / (optimize.Length + 0.0)) * 100.0) - 100).ToString("0.00 '%'"));
                     foreach (var func in module.mapMethods)
                     {
-                        uint srcaddr = (uint)func.Value.funcaddr;
-                        uint opaddr = addrConvTable[srcaddr];
+                        int srcaddr = func.Value.funcaddr;
+                        int opaddr = addrConvTable[srcaddr];
                         log.Log("func addr from " + srcaddr + "=>" + opaddr);
                     }
                     bytes = optimize;
